+++ src/com/intellij/idea/plugin/hybris/system/type/meta/TSMetaModelProcessor.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -17,47 +17,49 @@
  */
 package com.intellij.idea.plugin.hybris.system.type.meta
 
-import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
 import com.intellij.idea.plugin.hybris.system.type.meta.impl.TSMetaModelBuilder
 import com.intellij.idea.plugin.hybris.system.type.model.Items
-import com.intellij.idea.plugin.hybris.system.type.model.TypeGroup
 import com.intellij.idea.plugin.hybris.system.type.util.TSUtils
+import com.intellij.openapi.application.readAction
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 import com.intellij.psi.xml.XmlFile
 import com.intellij.util.xml.DomManager
-import java.util.stream.Collectors
+import kotlinx.coroutines.*
 
 @Service(Service.Level.PROJECT)
-class TSMetaModelProcessor(private val myProject: Project) {
+class TSMetaModelProcessor(myProject: Project) {
 
     private val myDomManager: DomManager = DomManager.getDomManager(myProject)
 
-    fun process(psiFile: PsiFile): TSMetaModel? {
-        psiFile.virtualFile ?: return null
-        val module = TSUtils.getModuleForFile(psiFile) ?: return null
+    suspend fun process(coroutineScope: CoroutineScope, psiFile: PsiFile): TSMetaModel? = coroutineScope {
+        psiFile.virtualFile ?: return@coroutineScope null
+        val module = TSUtils.getModuleForFile(psiFile) ?: return@coroutineScope null
         val custom = TSUtils.isCustomExtensionFile(psiFile)
         val rootWrapper = myDomManager.getFileElement(psiFile as XmlFile, Items::class.java)
 
-        rootWrapper ?: return null
+        rootWrapper ?: return@coroutineScope null
 
         val items = rootWrapper.rootElement
 
-        ProgressManager.getInstance().progressIndicator.text2 = HybrisI18NBundleUtils.message("hybris.ts.access.progress.subTitle.processing", psiFile.name)
+        val builder = TSMetaModelBuilder(module, psiFile, custom)
 
-        return TSMetaModelBuilder(module, psiFile, custom)
-            .withItemTypes(items.itemTypes.itemTypes)
-            .withItemTypes(items.itemTypes.typeGroups.stream()
-                .flatMap { tg: TypeGroup -> tg.itemTypes.stream() }
-                .collect(Collectors.toList()))
-            .withEnumTypes(items.enumTypes.enumTypes)
-            .withAtomicTypes(items.atomicTypes.atomicTypes)
-            .withCollectionTypes(items.collectionTypes.collectionTypes)
-            .withRelationTypes(items.relations.relations)
-            .withMapTypes(items.mapTypes.mapTypes)
-            .build()
+        val operations = listOf(
+            coroutineScope.async { readAction { builder.withItemTypes(items.itemTypes.itemTypes) } },
+            coroutineScope.async { readAction { builder.withItemTypes(items.itemTypes.typeGroups.flatMap { it.itemTypes }) } },
+            coroutineScope.async { readAction { builder.withEnumTypes(items.enumTypes.enumTypes) } },
+            coroutineScope.async { readAction { builder.withAtomicTypes(items.atomicTypes.atomicTypes) } },
+            coroutineScope.async { readAction { builder.withCollectionTypes(items.collectionTypes.collectionTypes) } },
+            coroutineScope.async { readAction { builder.withRelationTypes(items.relations.relations) } },
+            coroutineScope.async { readAction { builder.withMapTypes(items.mapTypes.mapTypes) } },
+        )
+
+        withContext(Dispatchers.IO) {
+            operations.awaitAll()
+        }
+
+        builder.build()
     }
 
     companion object {